requirements.txt
bottle==0.12.17
psycopg2==2.8.3
psycopg2-binary==2.8.3
pycrypto==2.6.1
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0